refactor: userAgent 분리
