+++ apps/web/src/utils/userAgent.ts
@@ -1,0 +1,15 @@
+export const isIos = () => {
+  const userAgent = navigator.userAgent;
+
+  return Boolean(userAgent.match(/iPhone|iPad|iPod/i));
+};
+
+export const isAndroid = () => {
+  const userAgent = navigator.userAgent;
+
+  return Boolean(userAgent.match(/Android/i));
+};
+
+export const isDesktop = () => {
+  return Boolean(!isAndroid() && !isIos());
+};
